kafka_consumer/requirements.txt
kafka-python
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
logging
prometheus-client